services/notification-service/requirements.txt
robyn[pydantic]
sqlalchemy
asyncpg
pydantic-settings
PyJWT
